[PA_1] Test Script implementaton and starter code
