+++ LE_6/graph.cpp
@@ -11,13 +11,9 @@ class Graph{
         list<pair<int,int>> *l;
     public:
         Graph(int size = 2){
-            n = size;
-            l = new list<pair<int,int>>[n];
 
         }
         void addEdge(int x, int y, int w){
-            l[x].push_back({y,w});
-            l[y].push_back({x,w});
         }
         void print(){
             for(int i = 0; i < n; i++){
@@ -32,49 +28,15 @@ class Graph{
         }
 
         int* dijkstra(){
-            int *dist = new int[n];
-            for(int i = 0; i < n; i++){
-                dist[i] = INT_MAX;
-            }
-            dist[0] = 0;
-            PriorityQueue <pair<int,int>> pq;
-            pq.pq_insert({0,0});
-
-            while(!pq.isPqEmpty()){
-                int currentNode = pq.pq_top().second;
-                int currentDist = pq.pq_top().first;
-                pq.pq_delete();
-                for(pair<int,int> i : l[currentNode]){
-                    if(currentDist + i.second < dist[i.first]){
-                        dist[i.first] = currentDist + i.second;
-                        pq.pq_insert({dist[i.first], i.first});
-                    }
-                }
-            }
-            return dist;
+            int *distances = new int[n];
+            
+            return distances;
         }
 };
 
 
-
-void testDijkstra(){
-    Graph g(5);
-    g.addEdge(0,1,1);
-    g.addEdge(1,2,5);
-    g.addEdge(2,3,7);
-    g.addEdge(3,4,3);
-    g.addEdge(1,4,2);
-    g.addEdge(0,3,5);
-    int* arr = g.dijkstra();
-    g.print();
-    cout<< "Dijkshtra"<<endl;
-    for(int i=0; i<5; i++){
-        cout << arr[i] << ", ";
-    }
-    cout<<endl;
-}
-
+#ifndef TEST
 int main(){
-    testDijkstra();
     return 0;
 }
+#endif
