+++ LE_6/solution/graph
