+++ LE_7/solution/Makefile
@@ -1,0 +1,2 @@
+mergesort: merge_sort.cpp
+	g++ merge_sort.cpp -o merge_sort
