+++ LE_8/solution/Makefile
@@ -1,0 +1,2 @@
+skiplist: skiplist.cpp
+	g++ skiplist.cpp -o skiplist
