+++ LE_9/solution/vector.csv
