+++ PA_1/Stack.cpp
@@ -7,7 +7,8 @@
 using namespace std;
 
 // The AbstractStack class defies the interface for a stack.
-// Implement all the virtual methods in this class. 
+// Create a class which extends AbstractStack like shown for class 'StackArrayDouble'.
+// Implement all the virtual methods in your class 'Stack' (not in AbstractStack). 
 
 class AbstractStack {
 private:
@@ -16,23 +17,11 @@ private:
 public:
     AbstractStack() {
     }
-
-    ~AbstractStack() {
-    }
-
+    // * Overload these function in your child class not here!
     virtual bool isEmpty() = 0;
     virtual int size() = 0;
-
-    // * Returns top element.
-    // * Throw an exception if the stack is empty.
     virtual int top() = 0;
-
-    // * Removes and returns top element.
-    // * Throw an exception if the stack is empty.
     virtual int pop() = 0;
-
-    // * Declared as pure virtual.
-    // * Overload this function in your child class!
     virtual void push (int e) = 0;
 };
 
@@ -41,41 +30,58 @@ public:
 #ifndef STACK_ARRAY_H
 #define STACK_ARRAY_H
 
-class Stack : public AbstractStack{
+class StackArrayDouble : public AbstractStack{
 private:
     int* data;
     int length;
     int topIndex;
 
 public:
-    Stack(){
+    StackArrayDouble(){
     }
 
-    ~Stack(){
+    ~StackArrayDouble(){
+        delete[] data;
     }
 
     bool isEmpty(){
-        return false;
     }
 
     int size(){
-        return 0;
+
     }
 
     int top(){
-        return 0;
+        
     }
 
     int pop(){
-        return 0;
+
     }
 
     void push(int e){
+
     }
 
 };
 #endif
 
+class StackArrayLinear : public AbstractStack{};
+class StackLinkedList : public AbstractStack{};
+
+int testStackPush(AbstractStack *s){
+    s->push(1);
+    s->push(2);
+    s->push(5);
+    if(s->size() != 3){
+        return 0;
+    }
+    return 10;
+}
+#ifndef TEST
 int main(){
     cout<<"Hello World"<<endl;
+    testStackPush(new StackArrayDouble());
+    return 0;
 }
+#endif
